experiments/adk/requirements.txt
memg-core>=0.7.2
google-adk
google-genai
python-dotenv
